#490 Refactor Refactor Refactor
