+++ src/backend/src/prisma-query-args/proposed-solutions.query-args.ts
@@ -1,0 +1,8 @@
+import { Prisma } from '@prisma/client';
+const proposedSolutionArgs = Prisma.validator<Prisma.Proposed_SolutionArgs>()({
+  include: {
+    createdBy: true
+  }
+});
+
+export default proposedSolutionArgs;
