+++ src/backend/src/prisma-query-args/scope-change-request.query-args.ts
@@ -1,0 +1,11 @@
+import { Prisma } from '@prisma/client';
+import proposedSolutionArgs from './proposed-solutions.query-args';
+
+const scopeCRArgs = Prisma.validator<Prisma.Scope_CRArgs>()({
+  include: {
+    why: true,
+    proposedSolutions: proposedSolutionArgs
+  }
+});
+
+export default scopeCRArgs;
