+++ src/backend/tests/change-requests.test.ts
@@ -1,4 +1,3 @@
-import request from 'supertest';
 import express from 'express';
 import prisma from '../src/prisma/prisma';
 import { batman, batmanSettings, superman, wonderwoman } from './test-data/users.test-data';
@@ -12,6 +11,8 @@ import {
 } from './test-data/change-requests.test-data';
 import { project1 } from './test-data/projects.test-data';
 import { CR_Type } from '@prisma/client';
+import ChangeRequestService from '../src/services/change-request.services';
+import { AccessDeniedException, HttpException, NotFoundException } from '../src/utils/errors.utils';
 
 const app = express();
 app.use(express.json());
@@ -24,39 +25,82 @@ describe('Change Requests', () => {
   describe('reviewChangeRequest', () => {
     test('reviewer not found errors', async () => {
       jest.spyOn(prisma.user, 'findUnique').mockResolvedValue(null);
-      const response = await request(app).post('/review').send(whipPayloadObject);
-
-      expect(response.body).toStrictEqual({ message: `User with id #${whipPayloadObject.reviewerId} not found` });
-      expect(response.status).toBe(404);
+      try {
+        await ChangeRequestService.reviewChangeRequest(
+          whipPayloadObject.reviewerId,
+          whipPayloadObject.crId,
+          whipPayloadObject.reviewNotes,
+          whipPayloadObject.accepted,
+          '1');
+      }
+      catch (e) {
+        // eslint-disable-next-line jest/no-conditional-expect
+        expect(e).toBeInstanceOf(NotFoundException);
+        // eslint-disable-next-line jest/no-conditional-expect
+        expect(e).toEqual(new NotFoundException('User', 1));
+      }
       expect(prisma.user.findUnique).toHaveBeenCalledTimes(1);
     });
 
     test('reviewer doesnt have access errors', async () => {
       jest.spyOn(prisma.user, 'findUnique').mockResolvedValueOnce(wonderwoman);
-      const response = await request(app)
-        .post('/review')
-        .send({ ...whipPayloadObject, accepted: false });
-      expect(response.body).toStrictEqual({ message: 'Access Denied' });
-      expect(response.status).toBe(403);
+
+      try {
+        await ChangeRequestService.reviewChangeRequest(
+          whipPayloadObject.reviewerId,
+          whipPayloadObject.crId,
+          whipPayloadObject.reviewNotes,
+          whipPayloadObject.accepted,
+          '1');
+      }
+      catch (e) {
+        // eslint-disable-next-line jest/no-conditional-expect
+        expect(e).toBeInstanceOf(AccessDeniedException);
+        // eslint-disable-next-line jest/no-conditional-expect
+        expect(e).toEqual(new AccessDeniedException
+          ());
+      }
       expect(prisma.user.findUnique).toHaveBeenCalledTimes(1);
     });
 
     test('change request not found errors', async () => {
       jest.spyOn(prisma.user, 'findUnique').mockResolvedValueOnce(batman);
       jest.spyOn(prisma.change_Request, 'findUnique').mockResolvedValueOnce(null);
-      const response = await request(app).post('/review').send(whipPayloadObject);
+      try {
+        await ChangeRequestService.reviewChangeRequest(
+          whipPayloadObject.reviewerId,
+          whipPayloadObject.crId,
+          whipPayloadObject.reviewNotes,
+          whipPayloadObject.accepted,
+          '1');
+      }
+      catch (e) {
+        // eslint-disable-next-line jest/no-conditional-expect
+        expect(e).toBeInstanceOf(NotFoundException);
+        // eslint-disable-next-line jest/no-conditional-expect
+        expect(e).toEqual(new NotFoundException('Change Request', whipPayloadObject.crId));
+      }
       expect(prisma.user.findUnique).toHaveBeenCalledTimes(1);
       expect(prisma.change_Request.findUnique).toHaveBeenCalledTimes(1);
-      expect(response.status).toBe(404);
-      expect(response.body).toStrictEqual({ message: `Change request with id #${whipPayloadObject.crId} not found` });
     });
 
     test('change request already reviewed', async () => {
       jest.spyOn(prisma.user, 'findUnique').mockResolvedValueOnce(batman);
       jest.spyOn(prisma.change_Request, 'findUnique').mockResolvedValueOnce(redesignWhip);
-      const response = await request(app).post('/review').send(whipPayloadObject);
-      expect(response.body).toStrictEqual({ message: 'This change request is already approved!' });
-      expect(response.status).toBe(400);
+      try {
+        await ChangeRequestService.reviewChangeRequest(
+          whipPayloadObject.reviewerId,
+          whipPayloadObject.crId,
+          whipPayloadObject.reviewNotes,
+          whipPayloadObject.accepted,
+          '1'
+        );
+      } catch (e) {
+        // eslint-disable-next-line jest/no-conditional-expect
+        expect(e).toBeInstanceOf(HttpException);
+        // eslint-disable-next-line jest/no-conditional-expect
+        expect(e).toEqual(new HttpException(400, 'This change request is already approved!'));
+      }
       expect(prisma.user.findUnique).toHaveBeenCalledTimes(1);
       expect(prisma.change_Request.findUnique).toHaveBeenCalledTimes(1);
     });
@@ -66,21 +110,41 @@ describe('Change Requests', () => {
       jest
         .spyOn(prisma.change_Request, 'findUnique')
         .mockResolvedValue({ ...redesignWhip, accepted: false, submitterId: 1 });
-      const response = await request(app).post('/review').send(whipPayloadObject);
+      try {
+        await ChangeRequestService.reviewChangeRequest(
+          whipPayloadObject.reviewerId,
+          whipPayloadObject.crId,
+          whipPayloadObject.reviewNotes,
+          whipPayloadObject.accepted,
+          '1'
+        );
+      } catch (e) {
+        // eslint-disable-next-line jest/no-conditional-expect
+        expect(e).toBeInstanceOf(AccessDeniedException);
+        // eslint-disable-next-line jest/no-conditional-expect
+        expect(e).toEqual(new AccessDeniedException());
+      }
       expect(prisma.user.findUnique).toHaveBeenCalledTimes(1);
       expect(prisma.change_Request.findUnique).toHaveBeenCalledTimes(1);
-      expect(response.body).toStrictEqual({ message: 'Access Denied' });
-      expect(response.status).toBe(403);
     });
 
     test('did not select proposed solution', async () => {
       jest.spyOn(prisma.user, 'findUnique').mockResolvedValue(batman);
       jest.spyOn(prisma.change_Request, 'findUnique').mockResolvedValueOnce({ ...redesignWhip, accepted: false });
-
-      const response = await request(app).post('/review').send(whipPayloadObject);
-
-      expect(response.body).toStrictEqual({ message: 'No proposed solution selected for scope change request' });
-      expect(response.status).toBe(400);
+      try {
+        await ChangeRequestService.reviewChangeRequest(
+          whipPayloadObject.reviewerId,
+          whipPayloadObject.crId,
+          whipPayloadObject.reviewNotes,
+          whipPayloadObject.accepted,
+          null
+        );
+      } catch (e) {
+        // eslint-disable-next-line jest/no-conditional-expect
+        expect(e).toBeInstanceOf(HttpException);
+        // eslint-disable-next-line jest/no-conditional-expect
+        expect(e).toEqual(new HttpException(400, 'No proposed solution selected for scope change request'));
+      }
       expect(prisma.user.findUnique).toHaveBeenCalledTimes(1);
       expect(prisma.change_Request.findUnique).toHaveBeenCalledTimes(1);
     });
@@ -88,15 +152,21 @@ describe('Change Requests', () => {
     test('proposed solution id not found', async () => {
       jest.spyOn(prisma.user, 'findUnique').mockResolvedValue(batman);
       jest.spyOn(prisma.change_Request, 'findUnique').mockResolvedValueOnce({ ...redesignWhip, accepted: null });
-      jest.spyOn(prisma.proposed_Solution, 'findUnique').mockResolvedValueOnce(null);
-      const psId = 100;
-      const response = await request(app)
-        .post('/review')
-        .send({ ...whipPayloadObject, psId: '100' });
-      expect(response.status).toBe(404);
-      expect(response.body).toStrictEqual({
-        message: `Proposed solution with id #${psId} not found for change request #${redesignWhip.crId}`
-      });
+      jest.spyOn(prisma.proposed_Solution, 'findUnique').mockResolvedValueOnce({...solutionToRedesignWhip, changeRequestId: 10});
+      try {
+        await ChangeRequestService.reviewChangeRequest(
+          whipPayloadObject.reviewerId,
+          whipPayloadObject.crId,
+          whipPayloadObject.reviewNotes,
+          whipPayloadObject.accepted,
+          '1'
+        );
+      } catch (e) {
+        // eslint-disable-next-line jest/no-conditional-expect
+        expect(e).toBeInstanceOf(HttpException);
+        // eslint-disable-next-line jest/no-conditional-expect
+        expect(e).toEqual(new HttpException(404, `Proposed solution with id #1 not found for change request #${redesignWhip.crId}`));
+      }
       expect(prisma.user.findUnique).toHaveBeenCalledTimes(1);
       expect(prisma.change_Request.findUnique).toHaveBeenCalledTimes(1);
       expect(prisma.proposed_Solution.findUnique).toHaveBeenCalledTimes(1);
@@ -112,13 +182,20 @@ describe('Change Requests', () => {
       const updatedSolution = { ...solutionToRedesignWhip, accepted: true };
       jest.spyOn(prisma.proposed_Solution, 'update').mockResolvedValueOnce(updatedSolution);
       jest.spyOn(prisma.project, 'findUnique').mockResolvedValueOnce(null);
-      const response = await request(app)
-        .post('/review')
-        .send({ ...whipPayloadObject, psId: '1' });
-      expect(response.status).toBe(404);
-      expect(response.body).toStrictEqual({
-        message: `Work package project not found`
-      });
+      try {
+        await ChangeRequestService.reviewChangeRequest(
+          whipPayloadObject.reviewerId,
+          whipPayloadObject.crId,
+          whipPayloadObject.reviewNotes,
+          whipPayloadObject.accepted,
+          '1'
+        );
+      } catch (e) {
+        // eslint-disable-next-line jest/no-conditional-expect
+        expect(e).toBeInstanceOf(NotFoundException);
+        // eslint-disable-next-line jest/no-conditional-expect
+        expect(e).toEqual(new NotFoundException('Project', 100));
+      }
       expect(prisma.user.findUnique).toHaveBeenCalledTimes(1);
       expect(prisma.change_Request.findUnique).toHaveBeenCalledTimes(1);
       expect(prisma.proposed_Solution.findUnique).toHaveBeenCalledTimes(1);
@@ -135,13 +212,14 @@ describe('Change Requests', () => {
       jest.spyOn(prisma.project, 'update').mockResolvedValueOnce(project1);
       jest.spyOn(prisma.user_Settings, 'findUnique').mockResolvedValueOnce(batmanSettings);
       jest.spyOn(prisma.change_Request, 'update').mockResolvedValueOnce({ ...redesignWhip, accepted: true });
-      const response = await request(app)
-        .post('/review')
-        .send({ ...whipPayloadObject, psId: '1' });
-      expect(response.status).toBe(200);
-      expect(response.body).toStrictEqual({
-        message: `Change request #${redesignWhip.crId} successfully reviewed.`
-      });
+      const response = await ChangeRequestService.reviewChangeRequest(
+        whipPayloadObject.reviewerId,
+        whipPayloadObject.crId,
+        whipPayloadObject.reviewNotes,
+        whipPayloadObject.accepted,
+        '1'
+      );
+      expect(response).toStrictEqual(redesignWhip.crId);
       expect(prisma.user_Settings.findUnique).toHaveBeenCalledTimes(1);
       expect(prisma.user.findUnique).toHaveBeenCalledTimes(1);
       expect(prisma.change_Request.findUnique).toHaveBeenCalledTimes(1);
@@ -188,13 +266,15 @@ describe('Change Requests', () => {
       jest.spyOn(prisma.user_Settings, 'findUnique').mockResolvedValueOnce(batmanSettings);
       jest.spyOn(prisma.project, 'update').mockResolvedValueOnce(project1);
       jest.spyOn(prisma.change_Request, 'update').mockResolvedValueOnce({ ...redesignWhip, accepted: true });
-      const response = await request(app)
-        .post('/review')
-        .send({ ...whipPayloadObject, psId: '1' });
-      expect(response.status).toBe(200);
-      expect(response.body).toStrictEqual({
-        message: `Change request #${redesignWhip.crId} successfully reviewed.`
-      });
+      const response = await ChangeRequestService.reviewChangeRequest(
+        whipPayloadObject.reviewerId,
+        whipPayloadObject.crId,
+        whipPayloadObject.reviewNotes,
+        whipPayloadObject.accepted,
+        '1'
+      );
+      expect(response).toStrictEqual(redesignWhip.crId
+      );
       expect(prisma.user.findUnique).toHaveBeenCalledTimes(1);
       expect(prisma.change_Request.findUnique).toHaveBeenCalledTimes(1);
       expect(prisma.proposed_Solution.findUnique).toHaveBeenCalledTimes(1);
@@ -227,56 +307,78 @@ describe('Change Requests', () => {
         .spyOn(prisma.change_Request, 'findUnique')
         .mockResolvedValue(uncheckedExpectedActivitiesCR);
 
-      const response = await request(app).post('/review').send(whipPayloadObject);
+      try {
+        await ChangeRequestService.reviewChangeRequest(
+          whipPayloadObject.reviewerId,
+          whipPayloadObject.crId,
+          whipPayloadObject.reviewNotes,
+          whipPayloadObject.accepted,
+          '1'
+        );
+      } catch (e) {
+        // eslint-disable-next-line jest/no-conditional-expect
+        expect(e).toBeInstanceOf(HttpException);
+        // eslint-disable-next-line jest/no-conditional-expect
+        expect(e).toEqual(new HttpException(400, 'Work Package has unchecked expected activities'));
+      }
       expect(prisma.user.findUnique).toHaveBeenCalledTimes(1);
       expect(prisma.change_Request.findUnique).toHaveBeenCalledTimes(1);
-      expect(response.body).toStrictEqual({ message: 'Work Package has unchecked expected activities' });
-      expect(response.status).toBe(400);
     });
-  });
-  test('Stage Gate CR has unchecked deliverables', async () => {
-    jest.spyOn(prisma.user, 'findUnique').mockResolvedValue(batman);
-    const uncheckedDeliverables = {
-      ...whipWorkPackage.deliverables[0],
-      userCheckedId: null,
-      dateTimeChecked: null
-    };
-    const uncheckedDeliverablesWP = { ...whipWorkPackage, deliverables: [uncheckedDeliverables] };
-    const uncheckedDeliverablesWBS = { ...redesignWhipWBSElement, workPackage: uncheckedDeliverablesWP };
-    const uncheckedDeliverablesCR = {
-      ...redesignWhip,
-      wbsElement: uncheckedDeliverablesWBS,
-      accepted: false,
-      type: CR_Type.STAGE_GATE,
-      scopeChangeRequest: null
-    };
-    jest
-      .spyOn(prisma.change_Request, 'findUnique')
-      .mockResolvedValue(uncheckedDeliverablesCR);
+    test('Stage Gate CR has unchecked deliverables', async () => {
+      jest.spyOn(prisma.user, 'findUnique').mockResolvedValue(batman);
+      const uncheckedDeliverables = {
+        ...whipWorkPackage.deliverables[0],
+        userCheckedId: null,
+        dateTimeChecked: null
+      };
+      const uncheckedDeliverablesWP = { ...whipWorkPackage, deliverables: [uncheckedDeliverables] };
+      const uncheckedDeliverablesWBS = { ...redesignWhipWBSElement, workPackage: uncheckedDeliverablesWP };
+      const uncheckedDeliverablesCR = {
+        ...redesignWhip,
+        wbsElement: uncheckedDeliverablesWBS,
+        accepted: false,
+        type: CR_Type.STAGE_GATE,
+        scopeChangeRequest: null
+      };
+      jest.spyOn(prisma.change_Request, 'findUnique').mockResolvedValue(uncheckedDeliverablesCR);
 
-    const response = await request(app).post('/review').send(whipPayloadObject);
-    expect(prisma.user.findUnique).toHaveBeenCalledTimes(1);
-    expect(prisma.change_Request.findUnique).toHaveBeenCalledTimes(1);
-    expect(response.body).toStrictEqual({ message: 'Work Package has unchecked deliverables' });
-    expect(response.status).toBe(400);
+      try {
+        await ChangeRequestService.reviewChangeRequest(
+          whipPayloadObject.reviewerId,
+          whipPayloadObject.crId,
+          whipPayloadObject.reviewNotes,
+          whipPayloadObject.accepted,
+          '1'
+        );
+      } catch (e) {
+        // eslint-disable-next-line jest/no-conditional-expect
+        expect(e).toBeInstanceOf(HttpException);
+        // eslint-disable-next-line jest/no-conditional-expect
+        expect(e).toEqual(new HttpException(400, 'Work Package has unchecked deliverables'));
+      }
+      expect(prisma.user.findUnique).toHaveBeenCalledTimes(1);
+      expect(prisma.change_Request.findUnique).toHaveBeenCalledTimes(1);
+    });
+    test('Stage Gate CR Succeeds', async () => {
+      jest.spyOn(prisma.user, 'findUnique').mockResolvedValue(batman);
+      const validStageGateCR = { ...redesignWhip, accepted: false, type: CR_Type.STAGE_GATE, scopeChangeRequest: null };
+      jest.spyOn(prisma.change_Request, 'findUnique').mockResolvedValue(validStageGateCR);
+      jest.spyOn(prisma.change_Request, 'update').mockResolvedValue({ ...redesignWhip, accepted: true });
+      jest.spyOn(prisma.work_Package, 'update').mockResolvedValue(whipWorkPackage);
+      jest.spyOn(prisma.user_Settings, 'findUnique').mockResolvedValue(batmanSettings);
+      const response = await ChangeRequestService.reviewChangeRequest(
+        whipPayloadObject.reviewerId,
+        whipPayloadObject.crId,
+        whipPayloadObject.reviewNotes,
+        whipPayloadObject.accepted,
+        null
+      );
+      expect(prisma.user.findUnique).toHaveBeenCalledTimes(1);
+      expect(prisma.change_Request.findUnique).toHaveBeenCalledTimes(1);
+      expect(prisma.change_Request.update).toHaveBeenCalledTimes(1);
+      expect(prisma.work_Package.update).toHaveBeenCalledTimes(1);
+      expect(prisma.user_Settings.findUnique).toHaveBeenCalledTimes(1);
+      expect(response).toStrictEqual(whipPayloadObject.crId);
   });
-
-  test('Stage Gate CR Succeeds', async () => {
-    jest.spyOn(prisma.user, 'findUnique').mockResolvedValue(batman);
-    const validStageGateCR = { ...redesignWhip, accepted: false, type: CR_Type.STAGE_GATE, scopeChangeRequest: null };
-    jest
-      .spyOn(prisma.change_Request, 'findUnique')
-      .mockResolvedValue(validStageGateCR);
-    jest.spyOn(prisma.change_Request, 'update').mockResolvedValue({ ...redesignWhip, accepted: true });
-    jest.spyOn(prisma.work_Package, 'update').mockResolvedValue(whipWorkPackage);
-    jest.spyOn(prisma.user_Settings, 'findUnique').mockResolvedValue(batmanSettings);
-    const response = await request(app).post('/review').send(whipPayloadObject);
-    expect(prisma.user.findUnique).toHaveBeenCalledTimes(1);
-    expect(prisma.change_Request.findUnique).toHaveBeenCalledTimes(1);
-    expect(prisma.change_Request.update).toHaveBeenCalledTimes(1);
-    expect(prisma.work_Package.update).toHaveBeenCalledTimes(1);
-    expect(prisma.user_Settings.findUnique).toHaveBeenCalledTimes(1);
-    expect(response.body).toStrictEqual({ message: 'Change request #2 successfully reviewed.' });
-    expect(response.status).toBe(200);
-  });
+});
 });
